+++ src/gvisor_sandbox/__init__.py
@@ -15,7 +15,7 @@ from .types import (
     UnsupportedConfiguration,
 )
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 __all__ = [
     "AwsSandbox",